enemy statok rendezése
van funkció az ellenségek erősítésére
minden damage és healt érték float

+++ Assets/Prefabs/Lövedékek/Playerbullet.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class PlayerBullet : MonoBehaviour
 {
-    public int damage = 1; // Amount of damage the bullet deals
+    public float damage = 1; // Amount of damage the bullet deals
 
     private void OnCollisionEnter2D(Collision2D collision)
     {
